+++ iOS10-HW22 IsakovAnton/CoreData.swift
@@ -1,6 +1,7 @@
 import Foundation
 import CoreData
 
+// Протокол для взаимодействия с вторым экраном
 protocol SecondScreenView: class {
     func displayProfileData(name: String, birthDate: Date, gender: Gender)
 }
@@ -59,6 +60,8 @@ class CoreDataManager {
     
     func saveProfileData(_ updatedProfileData: Profile) {
         let context = persistentContainer.viewContext
+        
+        // Сохраните контекст, чтобы сохранить изменения
         do {
             try context.save()
         } catch {
@@ -67,25 +70,16 @@ class CoreDataManager {
     }
 }
 
-class FirstScreenPresenter {
-    weak var view: FirstScreenView?
-    var coreDataManager: CoreDataManager
-    
-    init(view: FirstScreenView, coreDataManager: CoreDataManager) {
-        self.view = view
-        self.coreDataManager = coreDataManager
-    }
-    
-    func addDataToCoreData(text: String) {
-        coreDataManager.addData(text)
-        view?.updateTableView()
-    }
-    
-    func deleteDataFromCoreData(at index: Int) {
-        coreDataManager.deleteData(at: index)
-        view?.updateTableView()
-    }
-}
+
+
+
+
+
+
+
+
+
+
 
 class SecondScreenPresenter {
     weak var view: SecondScreenView?
@@ -106,3 +100,4 @@ class SecondScreenPresenter {
         coreDataManager.saveProfileData(updatedProfileData)
     }
 }
+
