+++ iOS10-HW22 IsakovAnton/MainScreen/Assembly/MainAssembly.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+class MainAssembly {
+    class func configureModule() -> FirstScreenViewController {
+        let view = FirstScreenViewController()
+        let presenter = FirstScreenPresenter()
+        view.presenter = presenter
+        presenter.view = view
+        return view
+    }
+}
