+++ iOS10-HW22 IsakovAnton/MainScreen/Model/CellFirstView.swift
